chore: move option file
